Fix index out of range
Replace underlying structure of the board, for a matrix (slice of
slice), since it's a most accurate data structure, subject to less
errors on index calculations.

First scenario solved. More tests remain to check all scenarios are
fixed, but the basic play for a 19x19 board works for the first play (a
new test verifies so).