Fix index out of range error
The length of the variable containing the size of the board has to be
incremented, in order to prevent this error for larger boards (19 and
on).

It used to be a ``uint8``, which fits well for 19, but not for 19 x 19 =
361, because in this case was truncating to 105. Therefore when trying
to index, it was failing because the array was of length 105, and for
products greater than that, it was out of the limits.

The solution was to change the type for a ``uint16``, that for a board
size of 19, works.

Fixes https://github.com/ami-go/baduk-engine/issues/2